requirements.txt
flask==3.0.3
twilio==9.4.4
python-dotenv==1.0.0
gunicorn==21.2.0
reportlab==4.3.1
pillow==11.1.0
chardet==5.2.0